# backend/internal/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: db-manager API
  version: 0.1.0
  description: |
    Self-hosted database control plane API. User routes require a Bearer JWT or
    API token (`mdcp_` prefix). Agent routes under `/agent/v1` use agent bearer
    tokens (`mdcpa_` prefix).
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    agentAuth:
      type: http
      scheme: bearer
      description: Agent bearer token (mdcpa_ prefix)
  parameters:
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
paths:
...
  /v1/tokens:
    get:
      tags: [Tokens]
      security: [{bearerAuth: []}]
      summary: List API tokens
      responses:
        "200":
          description: Tokens
    post:
      tags: [Tokens]
      security: [{bearerAuth: []}]
      summary: Create API token
      responses:
        "201":
          description: Token with raw secret
  /v1/tokens/{id}:
    delete:
      tags: [Tokens]
      security: [{bearerAuth: []}]
      summary: Revoke API token
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "204":
          description: Revoked